requirements.txt
python-dotenv
uvicorn
phonenumbers
fastapi
pydantic
telethon